% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fiveprime}
\alias{fiveprime}
\alias{threeprime}
\title{Get the 5' (or 3') coordinates from a GRanges object.}
\usage{
fiveprime(gr)

threeprime(gr)
}
\arguments{
\item{gr}{The GRanges on which to operate}
}
\description{
Get the 5' (or 3') coordinates from a GRanges object.
}
\seealso{
\code{start, end, width, strand}
}

