% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format.kMG}
\alias{format.kMG}
\title{Format numbers using metric prefixes (k for 1000, M for 1e6, G for 1e9, etc)}
\usage{
\method{format}{kMG}(x, fmt = "\%.0f\%s")
}
\arguments{
\item{x}{Vector of numbers to format}

\item{fmt}{The format to use for the part before the prefix (the magnitude of which has changed).
It must start with a number format (typically %f with some precision) and the string format
used for the prefix.}
}
\value{
a character vector
}
\description{
Format numbers using metric prefixes (k for 1000, M for 1e6, G for 1e9, etc)
}
\examples{
 format.kMG(10^rcauchy(10)) # may over or underrun
}
\seealso{
\code{\link{sprintf}}, \code{\link{formatC}}, \code{\link{prettyNum}}
}
